[TEST] Split `org.emmalanguage.testschema` as `[...].test.schema`.
