+++ emma-common/src/test/scala/org/emmalanguage/test/schema/Literature.scala
@@ -14,11 +14,22 @@
  * limitations under the License.
  */
 package org.emmalanguage
-package testschema
+package test.schema
 
-/** A simple domain for mathematical analysis. */
-object Math {
+object Literature {
 
-  case class Point(x: Int, y: Int)
+  case class Book(title: String, author: String)
 
+  case class Character(name: String, book: Book)
+
+  // ---------------------------------------------------------------------------
+  // The Hitchhiker's Guide to the Galaxy
+  // ---------------------------------------------------------------------------
+
+  val hhBook = Book("The Hitchhiker's Guide to the Galaxy", "Douglas Adams")
+
+  val hhCrts = Seq(
+    Character("Arthur Dent", hhBook),
+    Character("Zaphod Beeblebrox", hhBook),
+    Character("Prostetnic Vogon Jeltz", hhBook))
 }
