+++ emma-common/src/test/scala/org/emmalanguage/test/schema/Math.scala
@@ -14,22 +14,11 @@
  * limitations under the License.
  */
 package org.emmalanguage
-package testschema
+package test.schema
 
-object Literature {
+/** A simple domain for mathematical analysis. */
+object Math {
 
-  case class Book(title: String, author: String)
+  case class Point(x: Int, y: Int)
 
-  case class Character(name: String, book: Book)
-
-  // ---------------------------------------------------------------------------
-  // The Hitchhiker's Guide to the Galaxy
-  // ---------------------------------------------------------------------------
-
-  val hhBook = Book("The Hitchhiker's Guide to the Galaxy", "Douglas Adams")
-
-  val hhCrts = Seq(
-    Character("Arthur Dent", hhBook),
-    Character("Zaphod Beeblebrox", hhBook),
-    Character("Prostetnic Vogon Jeltz", hhBook))
 }
