+++ emma-language/src/test/scala/org/emmalanguage/compiler/lang/comprehension/ReDeSugarSpec.scala
@@ -19,7 +19,7 @@ package compiler.lang.comprehension
 import eu.stratosphere.emma.api.DataBag
 import compiler.BaseCompilerSpec
 import compiler.ir.ComprehensionSyntax._
-import testschema.Marketing._
+import test.schema.Marketing._
 
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
