+++ emma-language/src/test/scala/org/emmalanguage/compiler/lang/core/PrettyPrintSpec.scala
@@ -19,7 +19,7 @@ package compiler.lang.core
 import eu.stratosphere.emma.api.{DataBag, emma}
 import compiler.ir.ComprehensionSyntax._
 import compiler.BaseCompilerSpec
-import testschema.Marketing._
+import test.schema.Marketing._
 
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
